Conversions + Features
* Converted gpsmarkers
* Converted endconditions
* Converted nametags

gpsmarkers changes
* Changed the way group markers are created. There are now NATO symbols
based on the unit's size attached to the marker. These are defined in
the f3 setGroupID script.
* Changed the way vehicle markers are created. Now you define the
vehicle as well as what type it is in an array within the side's vehicle
array. Types are listed in the file.
* Moved all common edited values (vehicle arrays, ignore markers) to
settings.sqf file in that folder.

endconditions
* Split time checks from player alive checks.
* Added a settings.sqf file where a user can easily define which sides
should throw up a hint when they are eliminated.