Update JIP fix
Previous implementation was having issues with the time variable not
being synced properly and JIP clients weren't being caught by it. This
version runs on the server and shouldn't have that problem. Instead of
killing players it will just force them to activate the spectator
script. Had to make some modifications to the spectator script to ensure
it wasn't run twice. Also had to change the boundary script to look for
the new bc_isSpectator variable so out of bounds messages wouldn't be
displayed during JIP.