Add extra conditions to JIPfix
Should make sure things that are created on the fly don't get teleported
to the corner of the map. I didn't test it but previously it may have
been the case where creating a parachute for example would teleport the
parachute to the bottom left corner of the map and then push the player
who created it into spectator mode.